sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sequins-otel"
version = "0.1.0"
description = "OpenTelemetry distro for Sequins — zero-config observability for local development"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["opentelemetry", "sequins", "observability", "otel", "tracing"]
dependencies = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sequins_otel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
